// src/api/models/EventRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MayfieldGolfClub | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EventType } from './EventType';
import {
    EventTypeFromJSON,
    EventTypeFromJSONTyped,
    EventTypeToJSON,
    EventTypeToJSONTyped,
} from './EventType';

/**
 * 
 * @export
 * @interface EventRequest
 */
export interface EventRequest {
    /**
     * 
     * @type {string}
     * @memberof EventRequest
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof EventRequest
     */
    details: string;
    /**
     * 
     * @type {Date}
     * @memberof EventRequest
     */
    date: Date;
    /**
     * 
     * @type {EventType}
     * @memberof EventRequest
     */
    eventType: EventType;
}



/**
 * Check if a given object implements the EventRequest interface.
 */
export function instanceOfEventRequest(value: object): value is EventRequest {
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('details' in value) || value['details'] === undefined) return false;
    if (!('date' in value) || value['date'] === undefined) return false;
    if (!('eventType' in value) || value['eventType'] === undefined) return false;
    return true;
}

export function EventRequestFromJSON(json: any): EventRequest {
    return EventRequestFromJSONTyped(json, false);
}

export function EventRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): EventRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'],
        'details': json['details'],
        'date': (new Date(json['date'])),
        'eventType': EventTypeFromJSON(json['eventType']),
    };
}

export function EventRequestToJSON(json: any): EventRequest {
    return EventRequestToJSONTyped(json, false);
}

export function EventRequestToJSONTyped(value?: EventRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'details': value['details'],
        'date': ((value['date']).toISOString()),
        'eventType': EventTypeToJSON(value['eventType']),
    };
}