// src/api/models/EventResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MayfieldGolfClub | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EventType } from './EventType';
import {
    EventTypeFromJSON,
    EventTypeFromJSONTyped,
    EventTypeToJSON,
    EventTypeToJSONTyped,
} from './EventType';

/**
 * 
 * @export
 * @interface EventResponse
 */
export interface EventResponse {
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof EventResponse
     */
    details: string;
    /**
     * 
     * @type {Date}
     * @memberof EventResponse
     */
    date: Date;
    /**
     * 
     * @type {EventType}
     * @memberof EventResponse
     */
    eventType: EventType;
}



/**
 * Check if a given object implements the EventResponse interface.
 */
export function instanceOfEventResponse(value: object): value is EventResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('details' in value) || value['details'] === undefined) return false;
    if (!('date' in value) || value['date'] === undefined) return false;
    if (!('eventType' in value) || value['eventType'] === undefined) return false;
    return true;
}

export function EventResponseFromJSON(json: any): EventResponse {
    return EventResponseFromJSONTyped(json, false);
}

export function EventResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): EventResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'details': json['details'],
        'date': (new Date(json['date'])),
        'eventType': EventTypeFromJSON(json['eventType']),
    };
}

export function EventResponseToJSON(json: any): EventResponse {
    return EventResponseToJSONTyped(json, false);
}

export function EventResponseToJSONTyped(value?: EventResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'details': value['details'],
        'date': ((value['date']).toISOString()),
        'eventType': EventTypeToJSON(value['eventType']),
    };
}